Update dependency on net-ldap gem to 0.4.0
Version 0.4.0 adds support for UTF-8 but it was never pushed to rubygems.
Some folks created net-ldap-1 which is now outdated because it does not
contain code from [pull request #33][pull-request-33] which allows to search
using binary data. This means you will not be able to look up objects
by `objectid` in Active Directory.

That pull request was merged into [main repo][net-ldap-repo-merged-33].
Thanks to bundler we can easily use it. You just need to explicitly declare
dependencies on this commit and on git repository for net-ldap in the project
where you want to use omniauth-ldap like this:

    gem "net-ldap", :git => "https://github.com/ruby-ldap/ruby-net-ldap.git", :ref => "8a182675f4f777fa15bdffab64ead773e034b7d0"
    gem "omniauth-ldap", :git => "https://github.com/DimaD/omniauth-ldap.git", :branch => "update-net-ldap-version"

When situation is resolved and we have good up-to date
version of `net-ldap` we will not need this commit.

[pull-request-33]: https://github.com/ruby-ldap/ruby-net-ldap/pull/33 "Pull request #33 -- New filter type to allow searching of binary data"
[net-ldap-repo-merged-33]: https://github.com/ruby-ldap/ruby-net-ldap/commit/b6d9fbebfd6e52441ee31e2ff91f07d286901d5b "Main net-ldap repo commit with pull request #33"